forge--object-id: Never try to decode ID as base64 string
Nearly always ID actually is an integer.  In the one case where it
is a string (an assignable user for a Gitlab repository), then it
is a string that is *not* base64 encoded.

In the future we might run into base64 encoded strings again, in
which case we can add a new method or find another way to handle
that special case.  Until then stop trying to decode.

This is a breaking change that requires the database to be reset.